requirements.txt
discord.py
python-Levenshtein
